Added functionality for updating users.
This commit adds functionality for updating users through an API call. Allows the user to update the first name, last name, role, or org node of a user. Populates available options for roles and org nodes through combo boxes. Upon saving, user's information is updated in the table on the main screen.