Updated the way user objects were handled and displayed.
This commit changes the way that user objects were displayed throughout the application. Previously, methods such as  formatForCreation,  formatForListing, etc., were created to be used to format users in a specific type of way. This was removed to better handle passing data to different parts of the application. The application now uses getters methods to gather information about specific attributes.