Added login screen for the application.
This commit adds a login screen for the application. Rather than the account ID and the secret key being passed in the program like it was in previous commits, it is now input by the user. Upon a successful API request, the user is greeted with the table of users, otherwise, an error occurs.