Modified functions to pass in credentials.
This commit changes the way that credentials are passed around throughout the application. Previously, the credentials were stored as a variable rather than entered by a user. Because of the user now inputting their own credentials rather than them being stored while I tested the application, functions that interact with an API now require the credentials to be passed in and the application itself passes around the credentials where needed.