Fixed issue related to global styling within the user interface.
This commit fixes an issue that was effecting the user interface. Previously, parts of the user interface such as the search bar, the pagination buttons, and the bottom two input fields on the UpdateMagnementForm, were being styled by CSS on the Login page specifically. Since CSS operates on a global scale, there was unintended styling being done to the previously listed elements. I didn't originally realize that this concept existed, and because of it just did workarounds to make the user interface work. This commit re-writes many parts of the UI to now be styled without additional workarounds needing to be done to make it consistent throughout the application.