Added landing page to the application.
This commit adds a welcome/landing page to the application. Rather than just being a standalone login page that leads into the application, this new landing page will go over what the application does, how to find your account ID and secret key, and how to utilize the application. Currently this page just provides the bones to the page, and has a Get Started section which holds the login form.