Fixed issues with the create and update user form.
This commit refactors the UserManagementForm to remove duplicated code that was originally used to allow the form to act as both a create new user form, as well as an update user form. By storing the variations of the form within a variable, it allows the return statement to be more readable and easier to understand. As well, the prior commit to this form would not update the form fields properly whenever attempting to update a user. This commit fixes that issue.